Avoid compilation warnings in gnutls.c on 64-bit hosts.
 src/gnutls.c (emacs_gnutls_handshake): Avoid compiler warnings about
 "cast to pointer from integer of different size".

+++ src/gnutls.c
@@ -101,8 +101,8 @@ emacs_gnutls_handshake (struct Lisp_Process *proc)
          in.  For an Emacs process socket, infd and outfd are the
          same but we use this two-argument version for clarity.  */
       gnutls_transport_set_ptr2 (state,
-        			 (gnutls_transport_ptr_t) proc->infd,
-        			 (gnutls_transport_ptr_t) proc->outfd);
+        			 (gnutls_transport_ptr_t) (long) proc->infd,
+        			 (gnutls_transport_ptr_t) (long) proc->outfd);
 #endif
 
       proc->gnutls_initstage = GNUTLS_STAGE_TRANSPORT_POINTERS_SET;
